+++ packages/preview-middleware/src/base/flp.ts
@@ -35,6 +35,7 @@ import {
     getAppName,
     sanitizeRtaConfig
 } from './config';
+import { generateCdm } from './cdm';
 
 const DEFAULT_LIVERELOAD_PORT = 35729;
 
@@ -156,6 +157,9 @@ export class FlpSandbox {
             this.createTestSuite(this.test);
         }
 
+        if (this.flpConfig.enhancedHomePage) {
+            this.addCDMRoute();
+        }
         await this.addRoutesForAdditionalApps();
         this.logger.info(`Initialized for app ${id}`);
         this.logger.debug(`Configured apps: ${JSON.stringify(this.templateConfig.apps)}`);
@@ -455,11 +459,17 @@ export class FlpSandbox {
             }
         }
         if (!version) {
-            this.logger.error('Could not get UI5 version of application. Using 1.121.0 as fallback.');
-            version = '1.121.0';
+            this.logger.error('Could not get UI5 version of application. Using 1.130.0 as fallback.');
+            version = '1.130.0';
         }
         const [major, minor, patch] = version.split('.').map((versionPart) => parseInt(versionPart, 10));
         const label = version.split(/-(.*)/s)?.[1];
+
+        if ((major < 2 && minor < 123) || major >= 2 || label?.includes('legacy-free')) {
+            this.flpConfig.enhancedHomePage = this.templateConfig.enhancedHomePage = false;
+            this.logger.warn(`Feature enhancedHomePage disabled: UI5 version ${version} not supported.`);
+        }
+
         return {
             major,
             minor,
@@ -481,7 +491,8 @@ export class FlpSandbox {
             }.`
         );
         const filePrefix = ui5Version.major > 1 || ui5Version.label?.includes('legacy-free') ? '2' : '';
-        return readFileSync(join(__dirname, `../../templates/flp/sandbox${filePrefix}.html`), 'utf-8');
+        const template = this.flpConfig.enhancedHomePage ? 'cdm' : 'sandbox';
+        return readFileSync(join(__dirname, `../../templates/flp/${template}${filePrefix}.html`), 'utf-8');
     }
 
     /**
@@ -537,6 +548,20 @@ export class FlpSandbox {
                 );
             }
         }
+    }
+
+    /**
+     * Add routes for cdm.json required by FLP during bootstrapping via cdm.
+     *
+     */
+    private addCDMRoute(): void {
+        this.router.get(
+            '/cdm.json',
+            async (_req: EnhancedRequest | connect.IncomingMessage, res: Response | http.ServerResponse) => {
+                const json = generateCdm(this.templateConfig.apps);
+                this.sendResponse(res, 'application/json', 200, JSON.stringify(json));
+            }
+        );
     }
 
     /**