+++ packages/environment-check/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/environment-check
 
+## 0.17.94
+
+### Patch Changes
+
+-   Updated dependencies [ced5edf]
+    -   @sap-ux/axios-extension@1.19.2
+
+## 0.17.93
+
+### Patch Changes
+
+-   d873cbd: chore - Add repository metadata in package.json
+
 ## 0.17.92
 
 ### Patch Changes