+++ packages/odata-service-inquirer/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @sap-ux/odata-service-inquirer
 
+## 2.2.26
+
+### Patch Changes
+
+-   Updated dependencies [ced5edf]
+    -   @sap-ux/axios-extension@1.19.2
+    -   @sap-ux/inquirer-common@0.6.32
+
+## 2.2.25
+
+### Patch Changes
+
+-   Updated dependencies [ab81490]
+    -   @sap-ux/project-input-validator@0.5.0
+
+## 2.2.24
+
+### Patch Changes
+
+-   Updated dependencies [c69752b]
+    -   @sap-ux/project-input-validator@0.4.0
+
 ## 2.2.23
 
 ### Patch Changes