+++ packages/ui-service-inquirer/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @sap-ux/ui-service-inquirer
 
+## 0.0.8
+
+### Patch Changes
+
+-   Updated dependencies [ced5edf]
+    -   @sap-ux/axios-extension@1.19.2
+    -   @sap-ux/inquirer-common@0.6.32
+    -   @sap-ux/abap-deploy-config-inquirer@1.2.35
+    -   @sap-ux/odata-service-inquirer@2.2.26
+
+## 0.0.7
+
+### Patch Changes
+
+-   @sap-ux/odata-service-inquirer@2.2.25
+
+## 0.0.6
+
+### Patch Changes
+
+-   @sap-ux/odata-service-inquirer@2.2.24
+
+## 0.0.5
+
+### Patch Changes
+
+-   Updated dependencies [1bca2d8]
+    -   @sap-ux/abap-deploy-config-inquirer@1.2.34
+
 ## 0.0.4
 
 ### Patch Changes